requirements.txt
streamlit
snowflake-connector-python
pandas
numpy
python-dotenv
plotly
